+++ changelog.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Added support to read and show errors from the classic "get_settings_errors()" in DataView
+- Added support for MultiField and FieldTable for DataView
+- Added visible hints for developers on any error if developer modus is enabled in WordPress
+
+### Changed
+
+- Switch the DataView to script-based config for big configurations
+- Hide hidden sections in DataView
+- Optimized the usage of MultiField
+- Use identical styling for legends on each field in DataView
+- Extended the support for WP_List_Table in DataView
+- Optimized error handling for faulty JSON configurations
+- Optimized styling of tables in classic view
+
 ## [3.1.2] - 15.08.2026
 
 ### Changed